fix for hoisting issues
