+++ packages/graphql-codegen-cli/src/config.ts
@@ -1,7 +1,6 @@
 import { cosmiconfig, defaultLoaders } from 'cosmiconfig';
 import { resolve } from 'path';
-import { Types } from '@graphql-codegen/plugin-helpers';
-import { DetailedError } from '@graphql-codegen/core';
+import { DetailedError, Types } from '@graphql-codegen/plugin-helpers';
 import { Command } from 'commander';
 import { GraphQLConfig } from 'graphql-config';
 import { findAndLoadGraphQLConfig } from './graphql-config';
