+++ packages/graphql-codegen-cli/src/presets.ts
@@ -1,6 +1,5 @@
-import { Types } from '@graphql-codegen/plugin-helpers';
+import { DetailedError, Types } from '@graphql-codegen/plugin-helpers';
 import { resolve } from 'path';
-import { DetailedError } from '@graphql-codegen/core';
 
 export async function getPresetByName(name: string, loader: Types.PackageLoaderFn<{ preset?: Types.OutputPreset; default?: Types.OutputPreset }>): Promise<Types.OutputPreset> {
   const possibleNames = [`@graphql-codegen/${name}`, `@graphql-codegen/${name}-preset`, name];
