+++ packages/utils/plugins-helpers/src/index.ts
@@ -3,3 +3,4 @@ export * from './types';
 export * from './utils';
 export * from './helpers';
 export * from './federation';
+export * from './errors';
